priority done, proxy done

+++ Backend/mail/src/main/java/com/example/mail/Controller.java
@@ -15,7 +15,8 @@ public class Controller {
 
     @PostMapping("/getUser")
     public User getUser(@RequestBody String email){
-        service.setCurrentUser(service.getUser(email));
+//        service.setCurrentUser(service.getUser(email));
+        if(!service.setCurrentUser(service.getUser(email)))return null;
         return service.getUser(email);
     }
     @PostMapping("/signUp")
@@ -27,19 +28,21 @@ public class Controller {
     public boolean signIn(@RequestBody UserDto user) throws NoSuchAlgorithmException {
         return service.signIn(user);
     }
-    @GetMapping("/signOut")
-    public void signOut(){
-        service.signOut();
+    @PostMapping("/signOut")
+    public void signOut(@RequestBody String email){
+        service.signOut(email);
     }
 
     @PostMapping("/addMail")
     public SystemDto addMail(@RequestBody Mail mail){
-        service.setCurrentUser(service.getUser(mail.getFrom()));
+//        service.setCurrentUser(service.getUser(mail.getFrom()));
+        if(!service.setCurrentUser(service.getUser(mail.getFrom())))return null;
         return service.addMail(mail);
     }
     @PostMapping("/draftMail")
     public DraftFolder draftMail(@RequestBody Mail mail){
-        service.setCurrentUser(service.getUser(mail.getFrom()));
+//        service.setCurrentUser(service.getUser(mail.getFrom()));
+        if(!service.setCurrentUser(service.getUser(mail.getFrom())))return null;
         return service.draftMail(mail);
     }
 //    what to return??
@@ -49,76 +52,88 @@ public class Controller {
 //        return service.draftMail(mail);
 //    }
 
-//    @PostMapping("/trashMail")
+    //    @PostMapping("/trashMail")
 //    public TrashFolder trashMail(@RequestBody SystemDto systemDto){
 //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
 //        return service.trashMail(systemDto.getMail());
 //    }
     @PostMapping("/trashMail")
     public SystemDto trashMail(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+//        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         if(systemDto.getSource().equals("trash"))return service.deleteFromTrash(systemDto.getSourceMails());
         else return service.trashMail(systemDto.getSourceMails(),systemDto.getSource());
     }
     @PostMapping("/restoreTrash")
     public User restoreFromTrash(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+//        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.restoreFromTrash(systemDto.getSourceMails());
     }
-//    @PostMapping("deleteTrash")
+    //    @PostMapping("deleteTrash")
 //    public TrashFolder deleteFromTrash(@RequestBody SystemDto systemDto){
 //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
 //        return service.deleteFromTrash(systemDto.getSourceMails());
 //    }
     @PostMapping("/sentFolder")
     public SentFolder getSentFolder(@RequestBody String email){
-        service.setCurrentUser(service.getUser(email));
+        //        service.setCurrentUser(service.getUser(email));
+        if(!service.setCurrentUser(service.getUser(email)))return null;
         return service.currentUser.getSentFolder();
     }
     @PostMapping("/inboxFolder")
     public InboxFolder getInboxFolder(@RequestBody String email){
-        service.setCurrentUser(service.getUser(email));
+        //        service.setCurrentUser(service.getUser(email));
+        if(!service.setCurrentUser(service.getUser(email)))return null;
         return service.currentUser.getInboxFolder();
     }
     @PostMapping("/trashFolder")
     public TrashFolder getTrashFolder(@RequestBody String email){
-        service.setCurrentUser(service.getUser(email));
+        //        service.setCurrentUser(service.getUser(email));
+        if(!service.setCurrentUser(service.getUser(email)))return null;
         return service.currentUser.getTrashFolder();
     }
     @PostMapping("/search")
     public ArrayList<Mail> search(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.search(systemDto.getHashMap(), systemDto.getSourceMails());
     }
     @PostMapping("/filter")
     public ArrayList<Mail> filter(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.filter(systemDto.getHashMap(), systemDto.getSourceMails());
     }
     @PostMapping("/addContact")
     public ArrayList<Contact> addContact(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.addContact(systemDto.getContact());
     }
     @PostMapping("/editContact")
     public ArrayList<Contact> editContact(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.editContact(systemDto);
     }
     @PostMapping("/deleteContact")
     public ArrayList<Contact> deleteContact(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.deleteContact(systemDto);
     }
     //only need email of current user and source = search written bas
     @PostMapping("/searchContacts")
     public ArrayList<Contact> searchContacts(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.searchContacts(systemDto);
     }
     @PostMapping("sortContacts")
     public ArrayList<Contact> sortContacts(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.sortContacts();
     }
     /*
@@ -128,23 +143,33 @@ public class Controller {
     */
     @PostMapping("/createFolder")
     public ArrayList<CustomFolder> createFolder(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.createCustomFolder(systemDto.getSource());
     }
     @PostMapping("/renameFolder")
     public ArrayList<CustomFolder> renameFolder(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.renameFolder(systemDto.getSource(), systemDto.getDestination());
     }
     @PostMapping("/removeFolder")
     public ArrayList<CustomFolder> removeFolder(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.removeFolder(systemDto.getSource());
     }
     @PostMapping("/moveToFolder")
     public ArrayList<CustomFolder> moveToFolder(@RequestBody SystemDto systemDto){
-        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        //        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
         return service.addToCustom(systemDto.getDestination(), systemDto.getSourceMails());
+    }
+    @PostMapping("/defaultOrPriority")
+    public ArrayList<Mail> defaultOrPriority(@RequestBody SystemDto systemDto){
+//        service.setCurrentUser(service.getUser(systemDto.getEmail()));
+        if(!service.setCurrentUser(service.getUser(systemDto.getEmail())))return null;
+        return service.defaultOrPriority(systemDto);
     }
 //    @PostMapping("/restoreTrash")
 //    public TrashFolder restoreFromTrash(@RequestBody SystemDto systemDto){
